requirements.txt

numpy==1.24.4
coloredlogs==15.0.1
pause==0.3
pandas==1.1.5
python-dotenv==1.0.0
expiringdict==1.2.2

# images
utils==1.0.1
PyAutoGUI==0.9.54
imutils==0.5.4
matplotlib==3.3.4
scikit-learn==0.24.0

# minecraft agent environment
minecraft-data==3.20.0
pygetwindow~=0.0.9
pynput==1.7.6
mouse==0.7.1
opencv-python==4.8.1.78
tinycss==0.4
screeninfo==0.8.1

# torch
tensorboard==2.10.1
torch==2.0.1
torchvision==0.15.2
torchviz==0.0.2
# ai training metrics
wandb==0.16.0

# monai swin_unetr
monai==1.2.0
nibabel==3.1.1
einops==0.4.1
tensorboardX==2.6

# minio storage server
minio==7.1.15
urllib3==1.25.11
pillow~=9.4.0
numerize==0.12

# server with rest endpoints
fastapi==0.101.1
uvicorn[standard]==0.19.0
starlette==0.27.0
httpx==0.22.0
requests==2.27.1
fastapi_utils==0.2.1
python-multipart==0.0.6
aiofiles==23.2.1
